// packages/core/src/compiler/lexer/token.d.ts
export declare enum TokenKind {
    NumberLiteral = "NumberLiteral",
    HexLiteral = "HexLiteral",
    BinaryLiteral = "BinaryLiteral",
    BooleanLiteral = "BooleanLiteral",
    StringLiteral = "StringLiteral",
    Identifier = "Identifier",
    Const = "Const",
    Let = "Let",
    Function = "Function",
    Return = "Return",
    If = "If",
    Else = "Else",
    Export = "Export",
    NumberType = "NumberType",
    BooleanType = "BooleanType",
    Plus = "Plus",
    Minus = "Minus",
    Star = "Star",
    Tilde = "Tilde",
    Ampersand = "Ampersand",
    Pipe = "Pipe",
    Caret = "Caret",
    ShiftLeft = "ShiftLeft",
    ShiftRight = "ShiftRight",
    ArithmeticShiftRight = "ArithmeticShiftRight",
    StrictEqual = "StrictEqual",
    StrictNotEqual = "StrictNotEqual",
    GreaterThan = "GreaterThan",
    LessThan = "LessThan",
    GreaterThanOrEqual = "GreaterThanOrEqual",
    LessThanOrEqual = "LessThanOrEqual",
    Equals = "Equals",
    LeftParen = "LeftParen",
    RightParen = "RightParen",
    LeftBrace = "LeftBrace",
    RightBrace = "RightBrace",
    LeftBracket = "LeftBracket",
    RightBracket = "RightBracket",
    Colon = "Colon",
    Semicolon = "Semicolon",
    Comma = "Comma",
    Class = "Class",
    This = "This",
    Switch = "Switch",
    Case = "Case",
    Default = "Default",
    Break = "Break",
    Enum = "Enum",
    Private = "Private",
    Public = "Public",
    Extends = "Extends",
    New = "New",
    Async = "Async",
    While = "While",
    At = "At",
    Dot = "Dot",
    PlusPlus = "PlusPlus",
    MinusMinus = "MinusMinus",
    PlusEquals = "PlusEquals",
    MinusEquals = "MinusEquals",
    Not = "Not",
    EndOfFile = "EndOfFile"
}
export interface Token {
    readonly kind: TokenKind;
    readonly value: string;
    readonly line: number;
    readonly column: number;
}
export declare const KEYWORD_TOKEN_MAP: ReadonlyMap<string, TokenKind>;